Theoretically, this GUI system should work. However, Sponge has not yet implemented the CustomInventories and so there is no way to test it yet. I'm going to check out SpongeVanilla and see if there is something I can do, but until then this is the end of the GUI project. Next up, signs.
