+++ common/src/main/java/net/buycraft/plugin/gui/Node.java
@@ -1,0 +1,8 @@
+package net.buycraft.plugin.gui;
+
+/**
+ * Created by meyerzinn on 2/18/16.
+ */
+public class Node {
+
+}
