+++ sponge/src/main/java/net/buycraft/plugin/sponge/command/GuiCmd.java
@@ -1,0 +1,8 @@
+package net.buycraft.plugin.sponge.command;
+
+/**
+ * Created by meyerzinn on 2/19/16.
+ */
+public class GuiCmd {
+
+}
