--- sponge/src/main/java/net/buycraft/plugin/sponge/gui/PackageIcon.java
@@ -1,13 +1,0 @@
-package net.buycraft.plugin.sponge.gui;
-
-import org.spongepowered.api.entity.living.player.Player;
-
-/**
- * Created by meyerzinn on 2/17/16.
- */
-public class PackageIcon implements GuiIcon {
-
-    @Override public void onClick(Player clicker) {
-
-    }
-}
