+++ sponge/src/main/java/net/buycraft/plugin/sponge/gui/icons/GuiIcon.java
@@ -1,5 +1,6 @@
-package net.buycraft.plugin.sponge.gui;
+package net.buycraft.plugin.sponge.gui.icons;
 
+import net.buycraft.plugin.sponge.gui.GuiView;
 import org.spongepowered.api.entity.living.player.Player;
 import org.spongepowered.api.item.inventory.ItemStack;
 
@@ -8,6 +9,6 @@ import org.spongepowered.api.item.inventory.ItemStack;
  */
 public interface GuiIcon {
 
-    void onClick(Player clicker);
+    void onClick(GuiView view, Player clicker);
 
 }
