+++ sponge/src/main/java/net/buycraft/plugin/sponge/gui/icons/NavigationIcon.java
@@ -1,0 +1,8 @@
+package net.buycraft.plugin.sponge.gui.icons;
+
+/**
+ * Created by meyerzinn on 2/19/16.
+ */
+public class NavigationIcon {
+
+}
